README link for Resources + Stack Overflow tag
Further to [this comment](https://github.com/pyscript/pyscript/pull/507#issuecomment-1198808273) in a rejected PR from the main PyScript repo, as advised I'm adding the SO link here to the community related pyscript-collective repo.  As the other resources are also likely relevant, I've added a section for them + links at the same time.